pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "smartswitch-explorer"
version = "0.1.0"
description = "Desktop Smart Switch backup explorer and exporter"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PySide6>=6.7",
  "pycryptodome>=3.20",
]

[dependency-groups]
dev = ["pytest>=8.2"]

[tool.uv]
default-groups = ["dev"]

[project.scripts]
smartswitch-explorer = "gui.main:main"

[tool.setuptools]
packages = [
  "gui",
  "gui.ui",
  "smartswitch_core",
  "smartswitch_core.applications",
  "smartswitch_core.crypto",
  "smartswitch_core.messages",
]

[tool.setuptools.package-dir]
"gui" = "src/gui"
"gui.ui" = "src/gui/ui"
"smartswitch_core" = "lib/smartswitch_core"
"smartswitch_core.applications" = "lib/smartswitch_core/applications"
"smartswitch_core.crypto" = "lib/smartswitch_core/crypto"
"smartswitch_core.messages" = "lib/smartswitch_core/messages"

[tool.setuptools.package-data]
"gui" = ["assets/*.png", "translations/*.ts", "translations/*.qm"]

[tool.pytest.ini_options]
testpaths = ["tests"]
